消息子类型cmd改为byte类型

+++ jforgame-net/src/main/java/com/kingston/jforgame/socket/annotation/MessageMeta.java
@@ -6,6 +6,8 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
+import com.kingston.jforgame.socket.message.Message;
+
 /**
  * The annotation type is used in {@link Message com.kingston.net.Message}
  * to specify module and cmd of the given message.
@@ -18,6 +20,6 @@ public @interface MessageMeta {
 
 	short module() default 0;
 
-	short cmd() default 0;
+	byte cmd() default 0;
 
 }
